Hub Connection
The hub room is the first room that is created before regular procedural
generation occurs. The Hub will spawn in the middle of the left side of
the board and then a corridor will connect it to the rest of the cave.
The player will always spawn inside the Hub.