Basic Testing Scene and Item Panning
Added testing platform to replace procedural generation to make it
easier to test the switch mechanics, added a basic switch as well as a
switch that pans over to an item in the scene and back.

Also changed both the player controller and the camera controller so
that the player jump does an instantaneous velocity change and the
camera controller has a smooth damp applied to it.