Hub Room
Just changed some things in the BoardCreator and Room Script.. so the
player spawns in the first room that is created and that the room starts
generating at the top left. Corridor needs to be a bit revised so that
for the first few iteation, the corridors are not random.