Pressure Plate + Misc.
-Pressure Plate Switch created and defined.
-Mask Layers were defined and assigned to the following:
Player(PlayerPlaceHolder)
Switch(Switch Collision Trigger Area)
Block(Floor/wall/ceiling block)
-The PlayerPlaceHolder prefab was changed so as to not rotate (circle
collider) when moving.
-PlayerPlaceHolder prefab was given PlayerGround point so that the
pressure plates detect it standing on it.

The pressure plate features a rising/sinking block that will sink when
the player stands on it and rise when it gets off it. The switch will
activate when the block sinks down a certain depth, like a button. Can
be placed arbitrarily. As of now only the player can sink the button,
future updates may allow other obstacles to activate the switch.