Sliding music staff + Modified buttons
Files:
./NEW/SlidingMusicStaff.prefab
./NEW/BaseButton.cs
./NEW/MusicStaff.cs

Summary:
- The prefab should contain the entire setup. Drop it in the Hierarchy.
- BaseButton.cs is the button used for the Slide Up and Slide Down
functionalities. See below for details.
- MusicStaff.cs contains code to move the music note up and down the
staff.

Usage:
Press E while the player is on top of either buttons to change the note
(currently a dot)'s position on the staff (currently blank).
The Slide Up button moves the note up and wraps around to the bottom if
it reaches the top.
The Slide Down button moves the note down and wraps around to the top if
it reaches the bottom.

On buttons:
I have made an experimental button that contains a UnityEvent.

This event invokes any functions specified in the inspector when the
player presses the button (stand on it, press E).

With this, there's no need to make specific "Response" scripts. If an
object wants to react to a button, it just needs to have public
functions that the button itself can call.
In the Sliding Music Staff example, when a button is pressed, it calls
the Slide Up/Down functions in the Music Staff's script, which changes
the note's position.

Also, we have way too many button scripts with largely similar code. I
suggest deriving them from an abstract/base class to avoid
rewriting/pasting the basic button functionalities.