Text Box Button
I was able to create a text button that makes a dialogue box appear
when you enter “E”. And by continuing to press “E” the text will appear
letter by letter.

For the prefab with the attached script component, I had to create a
canvas, text, game object, and text file to go with it.