Procedural Generation Failsafe
Variables I used for testing:
All are under GameManager object
-Columns: 100
-Rows: 100
-Num Rooms Min: 10
-Num Rooms Max: 10
-Corridor Prec Chance: 100

Basically smaller amount of space with more rooms to trigger the fail
safe.

There is a chance that the BoardCreator will be unable to find a good
corridor/room placement. It goes into an infinite loop and the game will
not load. These additions will prevent that by simply reloading the
level after BoardCreator has tried a certain amount of times to find
good placement.

Sometimes rooms or corridors in thier respective arrays in the
SetTileValuesForCorridor and SetTileValuesForRooms funcions will have a
null object in them and will throw a NullReferenceException. Fixes this
also by reloading the level.