TeleportToHub
So i got the teleport button on the last room with a prefab as a
placement. I also have the prefab for the Hub room so that the player
will teleport to. However, i dont know if its a bug on unity or my
terrible coding, but OnCollisionStay2D does not work. that is the only
thing i am stuck at. People are saying that 5.3 Unity is buggy but if
you have time can you check?