Dead end corridors
This is not the prefab placement but the next step. Corridors will now
have a percent chance for a dead end corridor to divert from them.

Currently when a non-dead corridor is going one direction, the dead end
corridor will go in the next direction. For example: Say that the
non-dead corridor is going East, then the dead corridor will go South.

Functionality can be added in the future for it to randomly pick between
the two perpendicular directions. Like if a non-dead corridor is going
North, then the dead-end corridor could go either East or West.