Prevent overlapping corridors and corridors on rooms
All corridors will now not overlap with each other.
All corridors will now not overlap with rooms.

During generation sometimes it will look like a corridor and a room are
fused together. They are NOT overlapping. The corridor is right next to
the room and thus there are no block layers to separate the two. This
will be a task in the future.

Sometimes during generation a room in the rooms array is null. Added a
fail safe for this event.

-Enclosed the teleporter instantiation in a small if statement so that
it doesn't throw a null exception.