Randomly Spawn Codex Placeholders
At the end of a dead end corridor there is a chance for a codex
placeholder, currently just memes, to spawn.

Also changed the maxLength of the dead end corridors so that dead end
corridors will properly be clamped to the board. Before if a dead end
corridor spawned within the length or width of a room from the edge (i.e
20 blocks in from any board edge) the dead end corridor will simply not
spawn. Changed it so that dead end corridors can now run to the board
edge.