Set Unused Blocks to Inactive
During load time all the tiles on the board are set to be inactive.
During gameplay the tiles in the player's view will be active. As the
player moves around blocks will become active. There is a timer that
after a designated amount goes through board and resets or "refreshing"
the board so that any tile outside of the player's view is inactive.
This increases the performance on both the PC and Mobile build.

NECESSARY:
A new tag "WallTile" needs to be created and all the tiles in the Wall
Tiles array needs to have that tag applied to them.  I am not including
the tag and the updated wall tiles as this could cause merge conflicts.