Merge of Inactive Blocks
During load time all the tiles on the board are set to be inactive.
During gameplay the tiles in the player's view will be active. As the
player moves around blocks will become active. There is a timer that
after a designated amount goes through board and resets or "refreshing"
the board so that any tile outside of the player's view is inactive.
This increases the performance on both the PC and Mobile build.