Two Character conversations.
This code will take in a text file formatted like the SampleNPC.txt.
The @ will determine new lines, and the names will specify who spoke it
(and which List it should be stored in).

It will alternate colors depending on who is talking, then end when the
last person is done speaking.

1. Specify the two characters talking.
2. Format the text file so it looks like this
NPC:Hello Zimri.@
Zimri:Hello stranger.
2a. Name the person saying the line, then end the line with @
Unless the line is at the very end of the textfile, then do NOT include
the @
2b. DO NOT double up speakers one right after the other.  This will not
work.
i.e.  Zimri: blahblah@
Zimri: blasdasd@
NPC: You weren't supposed to do that.