LoadingScreenScene
The canvases of the title screen have been moved slightly due to
overlapping problems in not being able to click on the buttons.

The Play button no longer loads up the GenerationTestScene directly. It
loads up a seperate scene that will preload the GenerationTestScene and
play it when it is done loading.