Short Corridor Placement Retry and Randomize Dead End Corridor Direction w/ a few bug fixes
Copying and pasting code results in marking pretty much everything as a
change. My apologies.

This commit consists of two tasks

1) (Procedural) Retry corridor placement if corridors are too
short(3hrs)
2) (Procedural)Randomize Dead End Corridor Directions(5hrs)

Changes for 1): In BoardCreator.cs I made a public variable
minCorridorLength. After creating a corridor because the corridor is
clamped to the board and that gives the true length of the corridor so I
added a couple if statements. If the length of the corridor is less than
the minCorridorLength then it retries placement.

Changes for 2): When the SetupDeadEndCorridor in Corridor.csis called it
takes the direction of the corridor that the dead end corridor is
spawning from and randomly picks which direction the corridor will go
but not in the reverse direction because that defeats the purpose of the
dead end corridors(i.e. If the already placed corridor is going North
then a dead end corridor cannot go South). After picking a direction it
sets the values and correctly positions the corridor with respect to the
directions of the dead end and already placed corridors.

-Also updated comparison logic in doCorridorsOverlapRooms to cover the
bug where a west direction corridor overlapping a room on the bottom
edge of the room with part of the corridor inside the room and the other
part outside the room
-Fixed the comparison logic in the doRoomsOverlapCorridors with a
corridor going in the east direction and the room placed on the room in
a way that the corridor is cutting through it. Had a greater than
compare when a lesser than was what was needed.
-Changed greater than and lesser than comparison logic to greater than
or equal to and lesser than or equal to in doCorridorsOverlapCorridor
-Fixed a typo that caused a bug in doCorridorsOverlapRooms with a west
direction corridor completely cutting through a room. It checked against
the y position of the room when it should have checked against the x
position.