Fixed Mobile Controls
The swipe down platform works great. This update merely updates the
mobile controls to mirror the performance and feel of the computer
version.

-Fixed the spacing of the code blocks, makes analyzing the code easier.
Some blocks were indented too far.
-Implemented the ray casting for mobile.
-Before players would continuously fall through subsequent platforms.
Took the swipe down for phasing through platforms and put it on top of
the mobile controls block to fix this. Had to take variable swipeValue
and turn it into a class data member instead of a local function data
member to achieve this.