Vertical Rain Stack Puzzle and Glowing Object
Created a script to spawn platforms a distance from each other with
switches under them. The switches are represented by the smiley faces.
When the player jumps onto the platform, the platform will tilt 90
degrees or -90 degrees depending on which side is stepped on. I have
also included the task where it has the glowing object. The object will
glow until you interact with it. Pressing the E button while in contact
with the object will disable the glow.