Floor Crumbling
Created a new scene with floor crumbling stuff. I used my own blocks
prefabs because I didn't want to modify the other blocks that we use for
levels. In order for the blocks to drop, they need a rigidbody component
and isKinematic needs to be true so they stay floating. The Maracas
prefab (aka the screenshake thing) is attacked to the block through the
crumbling floor script.