Prefab for Puzzles
Even though it says simon says prefab as the branch this is for the task
on prefabbing the other puzzles. I will try to create a prefab for the
simon says puzzle. I couldn't find all of them and some puzzles are
missing floors. Oh and i think 2 or 3 are not puzzles but i just put
them in just in case.