Five String Puzzle
A total of 5 notes the player will need to hit in order to complete the
puzzle. Only one note will be displayed at a time. The player needs to
hit the visible note in order to activate the next note in the sequence.
Once all of the notes are hit, the door will open.