Added final puzzle gate
-Created a gate prefab. This prevents the player from entering the last
puzzle room in the cave until all previous puzzles are finished.
-Created a script that keeps track of the number of puzzles finished. If
the correct number of puzzles are done then the gate is destroyed.
-Went through all the puzzle controller scripts and added a line that
increments the counter by 1 when it is finished. Once the puzzle is
finished it needs to be deactivated otherwise it causes the number of
rooms done to continuously increase.
-Edited the BoardCreator script to spawn the  gate