Final Puzzle
The final puzzle. SUPER annoying to test. The left and right wall are
used to determine the movement of the platforms and spheres. The bells
don't have any scripts or sounds. Platforms don't move initially, you
must hit the start button to start the puzzle, which can also be used to
reset the puzzle if the player messes up.